receipt-processor-app/requirements.txt
flask==2.3.3
werkzeug==2.3.7
firebase-admin==6.2.0
python-dotenv==1.0.0
pillow==10.0.1
pypdf2==3.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3